Add new `3d_dft` feature to avoid `libc` dependency
