+++ Cargo.toml
@@ -49,6 +49,7 @@ approx = "0.4"
 [features]
 default = []
 dft = ["feos-dft", "petgraph"]
+3d_dft = ["dft", "feos-dft/3d_dft"]
 estimator = []
 association = []
 pcsaft = ["association"]
@@ -56,4 +57,4 @@ gc_pcsaft = ["association"]
 uvtheory = ["lazy_static"]
 pets = []
 python = ["pyo3", "numpy", "feos-core/python", "feos-dft?/python"]
-all_models = ["dft", "estimator", "pcsaft", "gc_pcsaft", "uvtheory", "pets"]
+all_models = ["estimator", "pcsaft", "gc_pcsaft", "uvtheory", "pets"]
