+++ feos-dft/src/adsorption/fea_potential.rs
@@ -1,9 +1,9 @@
+use super::pore3d::{calculate_distance2, evaluate_lj_potential};
 use crate::profile::{CUTOFF_RADIUS, MAX_POTENTIAL};
 use crate::Geometry;
 use feos_core::EosUnit;
 use gauss_quad::GaussLegendre;
 use ndarray::{Array1, Array2, Zip};
-use ndarray_stats::SummaryStatisticsExt;
 use quantity::{QuantityArray2, QuantityScalar};
 use std::f64::consts::PI;
 use std::usize;
@@ -114,7 +114,7 @@ pub fn calculate_fea_potential<U: EosUnit>(
                 let distance2 = calculate_distance2(point, &coordinates, system_size);
                 let potential_sum: f64 = (0..sigma_sf.len())
                     .map(|alpha| {
-                        mi * evaluate(
+                        mi * evaluate_lj_potential(
                             distance2[alpha],
                             sigma_sf[alpha],
                             epsilon_k_sf[alpha],
@@ -125,44 +125,8 @@ pub fn calculate_fea_potential<U: EosUnit>(
                 potential_2d[[i1, i2]] = (-potential_sum.min(MAX_POTENTIAL)).exp();
             }
         }
-        *f = potential_2d.weighted_sum(&weights).unwrap();
+        *f = (potential_2d * &weights).sum();
     });
 
     -temperature * potential.map(|p| (p / weights_sum).ln())
-}
-
-// -temperature * potential.map(|p| (p / weights_sum).ln())
-
-/// Evaluate LJ12-6 potential between solid site "alpha" and fluid segment
-fn evaluate(distance2: f64, sigma: f64, epsilon: f64, cutoff_radius2: f64) -> f64 {
-    let sigma_r = sigma.powi(2) / distance2;
-
-    let potential: f64 = if distance2 > cutoff_radius2 {
-        0.0
-    } else if distance2 == 0.0 {
-        f64::INFINITY
-    } else {
-        4.0 * epsilon * (sigma_r.powi(6) - sigma_r.powi(3))
-    };
-
-    potential
-}
-
-/// Evaluate the squared euclidian distance between a point and the coordinates of all solid atoms.
-fn calculate_distance2(
-    point: [f64; 3],
-    coordinates: &Array2<f64>,
-    system_size: [f64; 3],
-) -> Array1<f64> {
-    Array1::from_shape_fn(coordinates.ncols(), |i| {
-        let mut rx = coordinates[[0, i]] - point[0];
-        let mut ry = coordinates[[1, i]] - point[1];
-        let mut rz = coordinates[[2, i]] - point[2];
-
-        rx = rx - system_size[0] * (rx / system_size[0]).round();
-        ry = ry - system_size[1] * (ry / system_size[1]).round();
-        rz = rz - system_size[2] * (rz / system_size[2]).round();
-
-        rx.powi(2) + ry.powi(2) + rz.powi(2)
-    })
 }
